Boson operators

TensorKitTensors.BosonOperators.a_min — Function

a_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minmin — Function

a_minmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minplus — Function

a_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic right-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plus — Function

a_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusmin — Function

a_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic left-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusplus — Function

a_plusplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.boson_space — Method

boson_space([symmetry::Type{<:Sector}]; cutoff::Integer)

The local Hilbert space for a (truncated) bosonic system with a given symmetry and cutoff. Available symmetries are Trivial or U1Irrep.

TensorKitTensors.BosonOperators.number — Function

number([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
n([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic number operator, with a maximum of cutoff bosons per site.