Fermion operators

TensorKitTensors.FermionOperators.f_hopping — Function

f_hopping([elt::Type{<:Number}=ComplexF64])
f_hop([elt::Type{<:Number}=ComplexF64])

Return the two-body operator that describes a particle that hops between the first and the second site.

TensorKitTensors.FermionOperators.f_min_f_min — Function

f_min_f_min([elt::Type{<:Number}=ComplexF64])
f⁻f⁻([elt::Type{<:Number}=ComplexF64])

Return the two-body operator that annihilates a particle at the first and at the second site.

TensorKitTensors.FermionOperators.f_min_f_plus — Function

f_min_f_plus([elt::Type{<:Number}=ComplexF64])
f⁻f⁺([elt::Type{<:Number}=ComplexF64])

Return the two-body operator that annihilates a particle at the first site and creates a particle at the second.

TensorKitTensors.FermionOperators.f_num — Function

f_num([elt::Type{<:Number}=ComplexF64])
n([elt::Type{<:Number}=ComplexF64])

Return the one-body operator that counts the nunber of particles.

TensorKitTensors.FermionOperators.f_plus_f_min — Function

f_plus_f_min([elt::Type{<:Number}=ComplexF64])
f⁺f⁻([elt::Type{<:Number}=ComplexF64])

Return the two-body operator that creates a particle at the first site and annihilates a particle at the second.

TensorKitTensors.FermionOperators.f_plus_f_plus — Function

f_plus_f_plus([elt::Type{<:Number}=ComplexF64])
f⁺f⁺([elt::Type{<:Number}=ComplexF64])

Return the two-body operator that creates a particle at the first and at the second site.

TensorKitTensors.FermionOperators.fermion_space — Method

fermion_space()

Return the local hilbert space for a model of spinless fermions.