$\mathrm{SU}(N)$ Hubbard operators

TensorKitTensors.SUNHubbardOperators — Module

module SUNHubbardOperators

Module for the $SU(N)$ generalizations of the HubbardOperators operators.

Note

This module requires using SUNRepresentations in order to define the SUNIrrep-symmetric tensors.

TensorKitTensors.SUNHubbardOperators.e_double — Function

e_double([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nn([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of doubly occupied flavor pairs on a given site.

\[nn = ∑_{α < β} n_α n_β\]

TensorKitTensors.SUNHubbardOperators.e_hopping — Function

e_hopping([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
e_hop([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that describes a particle that hops between the first and the second site.

TensorKitTensors.SUNHubbardOperators.e_min_e_plus — Function

e_min_e_plus([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
e⁻e⁺([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that annihilates a particle at the first site and creates a particle at the second.

TensorKitTensors.SUNHubbardOperators.e_num — Function

e_num([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
n([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of particles across all particle flavors.

\[n = ∑_{α} n_α\]

TensorKitTensors.SUNHubbardOperators.e_plus_e_min — Function

e_plus_e_min([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
e⁺e⁻([elt::Type{<:Number}], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that creates a particle at the first site and annihilates a particle at the second.

TensorKitTensors.SUNHubbardOperators.hubbard_space — Method

hubbard_space(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; N::Integer = 3)

Return the local hilbert space for a Hubbard-type model with the given particle and spin symmetries. The basis is spanned by all possible fermionic color-like combinations of particles that satisfy the Pauli exclusion principle.

For example, for $SU(3)$ (N = 3), we have three colors (r, g and b), leading to the following eight possible states:

\[|0⟩, |r⟩, |g⟩, |b⟩, |rg⟩, |rb⟩, |gb⟩, |rgb⟩\]

Additionally, we can implement Trivial or U1Irrep symmetry for the particle number conservation.