TensorKitTensors

Documentation for TensorKitTensors.

TensorKitTensors.BosonOperators.a_min

TensorKitTensors.BosonOperators.a_minmin

TensorKitTensors.BosonOperators.a_minplus

TensorKitTensors.BosonOperators.a_plus

TensorKitTensors.BosonOperators.a_plusmin

TensorKitTensors.BosonOperators.a_plusplus

TensorKitTensors.BosonOperators.boson_space

TensorKitTensors.BosonOperators.number

TensorKitTensors.SpinOperators.S_exchange

TensorKitTensors.SpinOperators.S_min

TensorKitTensors.SpinOperators.S_minplus

TensorKitTensors.SpinOperators.S_plus

TensorKitTensors.SpinOperators.S_plusmin

TensorKitTensors.SpinOperators.S_x

TensorKitTensors.SpinOperators.S_xx

TensorKitTensors.SpinOperators.S_y

TensorKitTensors.SpinOperators.S_yy

TensorKitTensors.SpinOperators.S_z

TensorKitTensors.SpinOperators.S_zz

TensorKitTensors.SpinOperators.spin_space

TensorKitTensors.SpinOperators.spinmatrices

TensorKitTensors.SpinOperators.σʸ

TensorKitTensors.SpinOperators.σˣ

TensorKitTensors.SpinOperators.σᶻ

TensorKitTensors.SpinOperators.σ⁺

TensorKitTensors.SpinOperators.σ⁻

TensorKitTensors.SpinOperators.S_exchange — Function

S_exchange([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
SS([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin exchange operator.

TensorKitTensors.SpinOperators.S_min — Function

S_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁻([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin lowering operator. See also σ⁻

TensorKitTensors.SpinOperators.S_minplus — Function

S_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁻⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin lowering-raising operator.

TensorKitTensors.SpinOperators.S_plus — Function

S_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin raising operator. See also σ⁺

TensorKitTensors.SpinOperators.S_plusmin — Function

S_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁺⁻([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin raising-lowering operator.

TensorKitTensors.SpinOperators.S_x — Function

S_x([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sˣ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the x-axis. See also σˣ

TensorKitTensors.SpinOperators.S_xx — Function

S_xx([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sˣˣ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin x exchange operator.

TensorKitTensors.SpinOperators.S_y — Function

S_y([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sʸ([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the y-axis. See also σʸ

TensorKitTensors.SpinOperators.S_yy — Function

S_yy([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sʸʸ([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin y exchange operator.

TensorKitTensors.SpinOperators.S_z — Function

S_z([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sᶻ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the z-axis. See also σᶻ

TensorKitTensors.SpinOperators.S_zz — Function

S_zz([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sᶻᶻ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin z exchange operator.

TensorKitTensors.SpinOperators.spin_space — Method

spin_space([symmetry::Type{<:Sector}]; spin=1 // 2)

The local Hilbert space for a spin system with a given symmetry and spin. Available symmetries are Trivial, Z2Irrep, U1Irrep and SU2Irrep.

TensorKitTensors.SpinOperators.spinmatrices — Function

spinmatrices(spin [, eltype])

The spinmatrices according to Wikipedia.

TensorKitTensors.SpinOperators.σʸ — Method

Pauli y operator.

TensorKitTensors.SpinOperators.σˣ — Method

Pauli x operator.

TensorKitTensors.SpinOperators.σᶻ — Method

Pauli z operator.

TensorKitTensors.SpinOperators.σ⁺ — Method

Pauli raising operator.

TensorKitTensors.SpinOperators.σ⁻ — Method

Pauli minus operator.

TensorKitTensors.BosonOperators.a_min — Function

a_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minmin — Function

a_minmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minplus — Function

a_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic right-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plus — Function

a_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusmin — Function

a_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic left-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusplus — Function

a_plusplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.boson_space — Method

boson_space([symmetry::Type{<:Sector}]; cutoff::Integer)

The local Hilbert space for a (truncated) bosonic system with a given symmetry and cutoff. Available symmetries are Trivial or U1Irrep.

TensorKitTensors.BosonOperators.number — Function

number([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
n([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic number operator, with a maximum of cutoff bosons per site.