TensorKitTensors

Documentation for TensorKitTensors.

TensorKitTensors.BosonOperators.a_min

TensorKitTensors.BosonOperators.a_minmin

TensorKitTensors.BosonOperators.a_minplus

TensorKitTensors.BosonOperators.a_plus

TensorKitTensors.BosonOperators.a_plusmin

TensorKitTensors.BosonOperators.a_plusplus

TensorKitTensors.BosonOperators.boson_space

TensorKitTensors.BosonOperators.number

TensorKitTensors.HubbardOperators.c_min_c_plus

TensorKitTensors.HubbardOperators.c_num

TensorKitTensors.HubbardOperators.c_plus_c_min

TensorKitTensors.HubbardOperators.d_min_d_plus

TensorKitTensors.HubbardOperators.d_num

TensorKitTensors.HubbardOperators.d_plus_d_min

TensorKitTensors.HubbardOperators.hubbard_space

TensorKitTensors.HubbardOperators.u_min_u_plus

TensorKitTensors.HubbardOperators.u_num

TensorKitTensors.HubbardOperators.u_plus_u_min

TensorKitTensors.HubbardOperators.ud_num

TensorKitTensors.SpinOperators.S_exchange

TensorKitTensors.SpinOperators.S_min

TensorKitTensors.SpinOperators.S_minplus

TensorKitTensors.SpinOperators.S_plus

TensorKitTensors.SpinOperators.S_plusmin

TensorKitTensors.SpinOperators.S_x

TensorKitTensors.SpinOperators.S_xx

TensorKitTensors.SpinOperators.S_y

TensorKitTensors.SpinOperators.S_yy

TensorKitTensors.SpinOperators.S_z

TensorKitTensors.SpinOperators.S_zz

TensorKitTensors.SpinOperators.spin_space

TensorKitTensors.SpinOperators.spinmatrices

TensorKitTensors.SpinOperators.σʸ

TensorKitTensors.SpinOperators.σˣ

TensorKitTensors.SpinOperators.σᶻ

TensorKitTensors.SpinOperators.σ⁺

TensorKitTensors.SpinOperators.σ⁻

TensorKitTensors.TJOperators.S_exchange

TensorKitTensors.TJOperators.S_min

TensorKitTensors.TJOperators.S_minplus

TensorKitTensors.TJOperators.S_plus

TensorKitTensors.TJOperators.S_plusmin

TensorKitTensors.TJOperators.S_x

TensorKitTensors.TJOperators.S_y

TensorKitTensors.TJOperators.S_z

TensorKitTensors.TJOperators.c_min_c_plus

TensorKitTensors.TJOperators.c_num

TensorKitTensors.TJOperators.c_num_hole

TensorKitTensors.TJOperators.c_plus_c_min

TensorKitTensors.TJOperators.c_singlet

TensorKitTensors.TJOperators.d_min_d_plus

TensorKitTensors.TJOperators.d_min_u_min

TensorKitTensors.TJOperators.d_num

TensorKitTensors.TJOperators.d_plus_d_min

TensorKitTensors.TJOperators.tj_space

TensorKitTensors.TJOperators.u_min_d_min

TensorKitTensors.TJOperators.u_min_u_plus

TensorKitTensors.TJOperators.u_num

TensorKitTensors.TJOperators.u_plus_u_min

TensorKitTensors.SpinOperators.S_exchange — Function

S_exchange([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
SS([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin exchange operator.

TensorKitTensors.SpinOperators.S_min — Function

S_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁻([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin lowering operator. See also σ⁻

TensorKitTensors.SpinOperators.S_minplus — Function

S_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁻⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin lowering-raising operator.

TensorKitTensors.SpinOperators.S_plus — Function

S_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin raising operator. See also σ⁺

TensorKitTensors.SpinOperators.S_plusmin — Function

S_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
S⁺⁻([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin raising-lowering operator.

TensorKitTensors.SpinOperators.S_x — Function

S_x([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sˣ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the x-axis. See also σˣ

TensorKitTensors.SpinOperators.S_xx — Function

S_xx([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sˣˣ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin x exchange operator.

TensorKitTensors.SpinOperators.S_y — Function

S_y([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sʸ([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the y-axis. See also σʸ

TensorKitTensors.SpinOperators.S_yy — Function

S_yy([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sʸʸ([eltype::Type{<:Complex}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin y exchange operator.

TensorKitTensors.SpinOperators.S_z — Function

S_z([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sᶻ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin operator along the z-axis. See also σᶻ

TensorKitTensors.SpinOperators.S_zz — Function

S_zz([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)
Sᶻᶻ([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; spin=1 // 2)

The spin z exchange operator.

TensorKitTensors.SpinOperators.spin_space — Method

spin_space([symmetry::Type{<:Sector}]; spin=1 // 2)

The local Hilbert space for a spin system with a given symmetry and spin. Available symmetries are Trivial, Z2Irrep, U1Irrep and SU2Irrep.

TensorKitTensors.SpinOperators.spinmatrices — Function

spinmatrices(spin [, eltype])

The spinmatrices according to Wikipedia.

TensorKitTensors.SpinOperators.σʸ — Method

Pauli y operator.

TensorKitTensors.SpinOperators.σˣ — Method

Pauli x operator.

TensorKitTensors.SpinOperators.σᶻ — Method

Pauli z operator.

TensorKitTensors.SpinOperators.σ⁺ — Method

Pauli raising operator.

TensorKitTensors.SpinOperators.σ⁻ — Method

Pauli minus operator.

TensorKitTensors.BosonOperators.a_min — Function

a_min([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minmin — Function

a_minmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_minplus — Function

a_minplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
aa⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic right-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plus — Function

a_plus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusmin — Function

a_plusmin([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic left-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.a_plusplus — Function

a_plusplus([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
a⁺a⁺([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.boson_space — Method

boson_space([symmetry::Type{<:Sector}]; cutoff::Integer)

The local Hilbert space for a (truncated) bosonic system with a given symmetry and cutoff. Available symmetries are Trivial or U1Irrep.

TensorKitTensors.BosonOperators.number — Function

number([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
n([eltype::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic number operator, with a maximum of cutoff bosons per site.

TensorKitTensors.HubbardOperators.c_min_c_plus — Function

c_min_c_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
c⁻c⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that annihilates a particle at the first site and creates a particle at the second. This is the sum of u_min_u_plus and d_min_d_plus.

TensorKitTensors.HubbardOperators.c_num — Function

c_num([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
n([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of particles.

TensorKitTensors.HubbardOperators.c_plus_c_min — Function

c_plus_c_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
c⁺c⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that creates a particle at the first site and annihilates a particle at the second. This is the sum of u_plus_u_min and d_plus_d_min.

TensorKitTensors.HubbardOperators.d_min_d_plus — Function

d_min_d_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
d⁻d⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the Hermitian conjugate of d_plus_d_min, i.e. $(c†_{1,↓}, c_{2,↓})† = -c_{1,↓}, c†_{2,↓}$ (note the extra minus sign). It annihilates a spin-down particle at the first site and creates a spin-down particle at the second.

TensorKitTensors.HubbardOperators.d_num — Function

d_num([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜜ([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of spin-down particles.

TensorKitTensors.HubbardOperators.d_plus_d_min — Function

d_plus_d_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
d⁺d⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator $c†_{1,↓}, c_{2,↓}$ that creates a spin-down particle at the first site and annihilates a spin-down particle at the second.

TensorKitTensors.HubbardOperators.hubbard_space — Function

hubbard_space(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector})

Return the local hilbert space for a Hubbard-type model with the given particle and spin symmetries. The possible symmetries are Trivial, U1Irrep, and SU2Irrep, for both particle number and spin.

TensorKitTensors.HubbardOperators.u_min_u_plus — Function

u_min_u_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
u⁻u⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the Hermitian conjugate of u_plus_u_min, i.e. $(c†_{1,↑}, c_{2,↑})† = -c_{1,↑}, c†_{2,↑}$ (note the extra minus sign). It annihilates a spin-up particle at the first site and creates a spin-up particle at the second.

TensorKitTensors.HubbardOperators.u_num — Function

u_num([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜛ([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of spin-up particles.

TensorKitTensors.HubbardOperators.u_plus_u_min — Function

u_plus_u_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
u⁺d⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator $c†_{1,↑}, c_{2,↑}$ that creates a spin-up particle at the first site and annihilates a spin-up particle at the second.

TensorKitTensors.HubbardOperators.ud_num — Function

ud_num([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜛꜜ([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of doubly occupied sites.

TensorKitTensors.TJOperators.S_exchange — Function

S_exchange(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the spin exchange operator S⋅S.

TensorKitTensors.TJOperators.S_min — Function

S_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the spin-minus operator.

TensorKitTensors.TJOperators.S_minplus — Function

S_minplus(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator S⁻S⁺. The only nonzero matrix element corresponds to |↓↑⟩ <-- |↑↓⟩.

TensorKitTensors.TJOperators.S_plus — Function

S_plus(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the spin-plus operator.

TensorKitTensors.TJOperators.S_plusmin — Function

S_plusmin(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator S⁺S⁻. The only nonzero matrix element corresponds to |↑↓⟩ <-- |↓↑⟩.

TensorKitTensors.TJOperators.S_x — Function

S_x(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body spin-1/2 x-operator on the electrons.

TensorKitTensors.TJOperators.S_y — Function

S_y(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body spin-1/2 x-operator on the electrons (only defined for Trivial symmetry).

TensorKitTensors.TJOperators.S_z — Function

S_z(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body spin-1/2 z-operator on the electrons.

TensorKitTensors.TJOperators.c_min_c_plus — Function

c_min_c_plus(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator that annihilates a particle at the first site and creates a particle at the second. This is the sum of u_min_u_plus and d_min_d_plus.

TensorKitTensors.TJOperators.c_num — Function

c_num(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body operator that counts the number of particles.

TensorKitTensors.TJOperators.c_num_hole — Function

c_num_hole(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body operator that counts the number of holes.

TensorKitTensors.TJOperators.c_plus_c_min — Function

c_plus_c_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator that creates a particle at the first site and annihilates a particle at the second. This is the sum of u_plus_u_min and d_plus_d_min.

TensorKitTensors.TJOperators.c_singlet — Function

c_singlet(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body singlet operator $(e_{1,↓} e_{2,↑} - e_{1,↓} e_{2,↑}) / sqrt(2)$.

TensorKitTensors.TJOperators.d_min_d_plus — Function

d_min_d_plus(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the Hermitian conjugate of d_plus_d_min, i.e. $(e†_{1,↓}, e_{2,↓})† = -e_{1,↓}, e†_{2,↓}$ (note the extra minus sign). It annihilates a spin-down electron at the first site and creates a spin-down electron at the second. The only nonzero matrix element corresponds to |0↓⟩ <-- |↓0⟩.

TensorKitTensors.TJOperators.d_min_u_min — Function

d_min_u_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator $e_{1,↓} e_{2,↑}$ that annihilates a spin-down particle at the first site and a spin-up particle at the second site. The only nonzero matrix element corresponds to |00⟩ <-- |↓↑⟩.

TensorKitTensors.TJOperators.d_num — Function

d_num(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool=false)

Return the one-body operator that counts the number of spin-down electrons.

TensorKitTensors.TJOperators.d_plus_d_min — Function

d_plus_d_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator $e†_{1,↓}, e_{2,↓}$ that creates a spin-down electron at the first site and annihilates a spin-down electron at the second. The only nonzero matrix element corresponds to |↓0⟩ <-- |0↓⟩.

TensorKitTensors.TJOperators.tj_space — Function

tj_space(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the local hilbert space for a t-J-type model with the given particle and spin symmetries. The possible symmetries are

Particle number: Trivial, U1Irrep;

Spin: Trivial, U1Irrep, SU2Irrep.

Setting slave_fermion = true switches to the slave-fermion basis.

basis states for slave_fermion = false: |0⟩ = |vac⟩ (vacuum), |↑⟩ = (c↑)†|vac⟩, |↓⟩ = (c↓)†|vac⟩

basis states for slave_fermion = true: (cσ = h† bσ; holon h is fermionic, spinon b_σ is bosonic): |0⟩ = h†|vac⟩, |↑⟩ = (b↑)†|vac⟩, |↓⟩ = (b↓)†|vac⟩

TensorKitTensors.TJOperators.u_min_d_min — Function

u_min_d_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator $e_{1,↑} e_{2,↓}$ that annihilates a spin-up particle at the first site and a spin-down particle at the second site. The only nonzero matrix element corresponds to |00⟩ <-- |↑↓⟩.

TensorKitTensors.TJOperators.u_min_u_plus — Function

u_min_u_plus(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the Hermitian conjugate of u_plus_u_min, i.e. $(e†_{1,↑}, e_{2,↑})† = -e_{1,↑}, e†_{2,↑}$ (note the extra minus sign). It annihilates a spin-up electron at the first site and creates a spin-up electron at the second. The only nonzero matrix element corresponds to |0↑⟩ <-- |↑0⟩.

TensorKitTensors.TJOperators.u_num — Function

u_num(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the one-body operator that counts the number of spin-up electrons.

TensorKitTensors.TJOperators.u_plus_u_min — Function

u_plus_u_min(T, particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector}; slave_fermion::Bool = false)

Return the two-body operator $e†_{1,↑}, e_{2,↑}$ that creates a spin-up electron at the first site and annihilates a spin-up electron at the second. The only nonzero matrix element corresponds to |↑0⟩ <-- |0↑⟩.