Boson operators

TensorKitTensors.BosonOperators.b_hopping — Function

b_hopping([elt::Type{<:Number}])
b_hop([elt::Type{<:Number}])

Return the two-body operator that describes a particle that hops between the first and the second site.

TensorKitTensors.BosonOperators.b_min — Function

b_min([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁻([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_min_b_min — Function

b_min_b_min([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁻b⁻([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-annihilation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_min_b_plus — Function

b_min_b_plus([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁻b⁺([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic right-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_num — Function

b_num([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
n([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic number operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_plus — Function

b_plus([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁺([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_plus_b_min — Function

b_plus_b_min([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁺b⁻([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic left-hopping operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.b_plus_b_plus — Function

b_plus_b_plus([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)
b⁺b⁺([elt::Type{<:Number}], [symmetry::Type{<:Sector}]; cutoff::Integer)

The truncated bosonic pair-creation operator, with a maximum of cutoff bosons per site.

TensorKitTensors.BosonOperators.boson_space — Method

boson_space([symmetry::Type{<:Sector}]; cutoff::Integer)

The local Hilbert space for a (truncated) bosonic system with a given symmetry and cutoff. Available symmetries are Trivial or U1Irrep.