Hubbard operators

TensorKitTensors.HubbardOperators.c_min_c_plus — Function

c_min_c_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
c⁻c⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that annihilates a particle at the first site and creates a particle at the second. This is the sum of u_min_u_plus and d_min_d_plus.

TensorKitTensors.HubbardOperators.c_num — Function

c_num([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
n([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of particles.

TensorKitTensors.HubbardOperators.c_plus_c_min — Function

c_plus_c_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
c⁺c⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator that creates a particle at the first site and annihilates a particle at the second. This is the sum of u_plus_u_min and d_plus_d_min.

TensorKitTensors.HubbardOperators.d_min_d_plus — Function

d_min_d_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
d⁻d⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the Hermitian conjugate of d_plus_d_min, i.e. $(c†_{1,↓}, c_{2,↓})† = -c_{1,↓}, c†_{2,↓}$ (note the extra minus sign). It annihilates a spin-down particle at the first site and creates a spin-down particle at the second.

TensorKitTensors.HubbardOperators.d_num — Function

d_num([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜜ([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of spin-down particles.

TensorKitTensors.HubbardOperators.d_plus_d_min — Function

d_plus_d_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
d⁺d⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator $c†_{1,↓}, c_{2,↓}$ that creates a spin-down particle at the first site and annihilates a spin-down particle at the second.

TensorKitTensors.HubbardOperators.hubbard_space — Function

hubbard_space(particle_symmetry::Type{<:Sector}, spin_symmetry::Type{<:Sector})

Return the local hilbert space for a Hubbard-type model with the given particle and spin symmetries. The possible symmetries are Trivial, U1Irrep, and SU2Irrep, for both particle number and spin.

TensorKitTensors.HubbardOperators.u_min_u_plus — Function

u_min_u_plus([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
u⁻u⁺([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the Hermitian conjugate of u_plus_u_min, i.e. $(c†_{1,↑}, c_{2,↑})† = -c_{1,↑}, c†_{2,↑}$ (note the extra minus sign). It annihilates a spin-up particle at the first site and creates a spin-up particle at the second.

TensorKitTensors.HubbardOperators.u_num — Function

u_num([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜛ([particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of spin-up particles.

TensorKitTensors.HubbardOperators.u_plus_u_min — Function

u_plus_u_min([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
u⁺d⁻([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the two-body operator $c†_{1,↑}, c_{2,↑}$ that creates a spin-up particle at the first site and annihilates a spin-up particle at the second.

TensorKitTensors.HubbardOperators.ud_num — Function

ud_num([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])
nꜛꜜ([T], [particle_symmetry::Type{<:Sector}], [spin_symmetry::Type{<:Sector}])

Return the one-body operator that counts the number of doubly occupied sites.